Rename dir for zod schema in topic request
