+++ coral/src/app/features/topics/request/components/AdvancedConfiguration.tsx
@@ -10,7 +10,7 @@ import { Position, editor } from "monaco-editor";
 import { useEffect, useRef } from "react";
 import { getTopicAdvancedConfigOptions } from "src/domain/topic/topic-api";
 import { TopicAdvancedConfigurationOptions } from "src/domain/topic/topic-types";
-import { Schema } from "src/app/features/topics/request/schemas/topic-request-form";
+import { Schema } from "src/app/features/topics/request/form-schemas/topic-request-form";
 import { BorderBox, Box, Flexbox, Typography } from "@aivenio/aquarium";
 
 type Props = {
