+++ coral/src/app/features/topics/request/components/SelectOrNumberInput.tsx
@@ -1,6 +1,6 @@
 import isNumber from "lodash/isNumber";
 import { NativeSelect, NumberInput } from "src/app/components/Form";
-import { Schema } from "src/app/features/topics/request/schemas/topic-request-form";
+import { Schema } from "src/app/features/topics/request/form-schemas/topic-request-form";
 
 type Props = {
   label: string;
